+++ packages/geoprocessing/scripts/base/datasources/genVectorConfig.ts
@@ -2,51 +2,46 @@ import path from "path";
 import {
   ImportVectorDatasourceOptions,
   ImportVectorDatasourceConfig,
+  VectorDatasource,
 } from "../../../src/types";
 import { datasourceConfig } from "../../../src/datasources";
-import { ProjectClientBase } from "../../../src";
+import { ProjectClientBase, hasOwnProperty } from "../../../src";
 
-/** Takes import options and creates full import config */
+/** Takes vector import options or existing datasource and creates full import config */
 export function genVectorConfig<C extends ProjectClientBase>(
   projectClient: C,
-  options: ImportVectorDatasourceOptions,
+  options: ImportVectorDatasourceOptions | VectorDatasource,
   newDstPath?: string
 ): ImportVectorDatasourceConfig {
-  let {
-    geo_type,
-    src,
-    datasourceId,
-    propertiesToKeep = [],
-    classKeys,
-    layerName,
-    formats = datasourceConfig.importDefaultVectorFormats,
-    explodeMulti,
-    precalc,
-    propertyFilter,
-    bboxFilter,
-  } = options;
+  const src =
+    hasOwnProperty(options, "src") && typeof options.src === "string"
+      ? options.src
+      : "";
 
-  if (!layerName)
-    layerName = path.basename(src, "." + path.basename(src).split(".").pop());
+  const explodeMulti =
+    hasOwnProperty(options, "explodeMulti") &&
+    typeof options.explodeMulti === "boolean"
+      ? options.explodeMulti
+      : false;
 
   // merge to ensure keep at least classKeys
-  propertiesToKeep = Array.from(new Set(propertiesToKeep.concat(classKeys)));
+  const propertiesToKeep =
+    hasOwnProperty(options, "propertiesToKeep") &&
+    Array.isArray(options.propertiesToKeep)
+      ? Array.from(new Set(options.propertiesToKeep.concat(options.classKeys)))
+      : [];
 
   const config: ImportVectorDatasourceConfig = {
-    geo_type,
+    ...options,
     src,
     dstPath: newDstPath || datasourceConfig.defaultDstPath,
     propertiesToKeep,
-    classKeys,
-    layerName,
-    datasourceId,
+    layerName:
+      options.layerName ||
+      path.basename(src, "." + path.basename(src).split(".").pop()),
     package: projectClient.package,
     gp: projectClient.geoprocessing,
-    formats,
     explodeMulti,
-    precalc,
-    propertyFilter,
-    bboxFilter,
   };
 
   return config;